move compose code
