+++ app/src/main/java/dev/leonlatsch/photok/cgallery/ui/GalleryViewModel.kt
@@ -19,12 +19,8 @@ package dev.leonlatsch.photok.cgallery.ui
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.leonlatsch.photok.cgallery.data.EncryptedImageFetcherFactory
 import dev.leonlatsch.photok.model.repositories.PhotoRepository
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import javax.inject.Inject
