+++ app/src/main/java/dev/leonlatsch/photok/imageloading/compose/EncryptedImageFetcher.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.cgallery.data
+package dev.leonlatsch.photok.imageloading.compose
 
 import android.content.Context
 import coil.decode.DataSource
@@ -23,11 +23,8 @@ import coil.fetch.FetchResult
 import coil.fetch.Fetcher
 import coil.fetch.SourceResult
 import dagger.hilt.android.qualifiers.ApplicationContext
-import dev.leonlatsch.photok.cgallery.data.model.EncryptedImageRequestData
-import dev.leonlatsch.photok.model.database.entity.Photo
+import dev.leonlatsch.photok.imageloading.compose.model.EncryptedImageRequestData
 import dev.leonlatsch.photok.model.io.EncryptedStorageManager
-import dev.leonlatsch.photok.model.repositories.PhotoRepository
-import dev.leonlatsch.photok.security.EncryptionManager
 import okio.buffer
 import okio.source
